refactor: broke out gameServer.js
